+++ tybotRoute/tiledeskChatbotPlugs/directives/DirAskGPTV2.js
@@ -62,8 +62,8 @@ class DirAskGPTV2 {
     let temperature;
     let max_tokens;
     let top_k;
-    let default_context = "Answer if and ONLY if the answer is contained in the context provided. If the answer is not contained in the context provided ALWAYS answer with <NOANS>\n{context}"
-    
+    let default_context = "You are an helpful assistant for question-answering tasks.\nUse ONLY the following pieces of retrieved context to answer the question.\nIf you don't know the answer, just say that you don't know.\nIf none of the retrieved context answer the question, add this word to the end <NOANS>\n\n{context}";
+
     let source = null;
 
     if (!action.question || action.question === '') {