+++ tybotRoute/CHANGELOG.md
@@ -7,6 +7,7 @@ available on:
 
 # v0.2.84
 - Fixed "guest#" bug
+- Fixed default_context in AskGPTv2
 
 # v0.2.83
 - Added support for advanced context in AskGPTv2